286 internalise K Sets (#292)
Fixes #286 

An internal K Set is a structure `KSet [Term] (Maybe Term)` with a list
of concrete set elements (sorted by hash but _not deduplicated_ ) and an
optional rest term (of compatible sort, which cannot be `KSet` by
construction).
If a set is constructed using set concatenation of several sets
containing such opaque rest terms, the concatenation is pushed inwards
(i.e., all concrete elements are gathered in the sorted list, and the
optional rest is the set concatenation of the opaque rest terms).

No attempt is made to match or unify these internal structures (yet). 

Set concatenation is partial as concatenating two sets with common
elements results in an undefined term. No attempt is made to check for
common/duplicate elements when internalising a set; the internalisation
retains those common elements as duplicates in the list of concrete
elements.

---------

Co-authored-by: Georgy Lukyanov <georgiylukjanov@gmail.com>