Add `Fallback` to `LogEntry` (#3657)
This PR modifies the `LogEntry` type in `kore-rpc-types` in two ways:
* add `origianlTerm` into `Rewrite` log entry constructor (and emit
Nothing for now in Kore to retain the current behavior. This change is
not necessary at the moment, but we may need to log these terms in
future.
* Add `Fallback` constructor to `LogEntry` to enable `kore-rpc-booster`
to emit a trace when it falls back from Booster to Kore.