Enable configuration of Z3 satisfiability tactics (#3661)
Fixes
[#3660](https://github.com/runtimeverification/haskell-backend/issues/3660)

This PR: 
* adds the `--smt-tactic <s-expression>` command line option [Z3
tactic](https://microsoft.github.io/z3guide/docs/strategies/tactics/);
* If the option is not supplied, the default tactic `(check-sat)` will
be used;
* The `execute` RPC integration tests are now run with a customized
tactic (and also without, as prior) to ensure that it is picked up and
the behavior is the same.

If a tactic is provided, it will be used to check satisfiability of
rewrite/simplification rules' side conditions. Tactics are specified as
SMT-LIB2 s-expression, but are Z3-specific, i.e. other solvers may not
support them.