Support multi-ary \and/\or in Haskell backend (#3666)
Part of #3542 

Essentially, we add support for \and and \or with 0 or more children.
The parser automatically desugars 0 or 1 children, so the AST will
always have at least 2 children. We modify the Internal modules to use a
new class, BinaryAnd and BinaryOr, which preserves the old
representation. Externalization and internalization then converts
between the two formats.

Most of this code is just needed to separate the `And` in
`Kore.Internal` from the `And` in `Kore.Syntax`. The actual changes are
relatively small.

### Scope:

### Estimate:

---

###### Review checklist

The author performs the actions on the checklist. The reviewer evaluates
the work and checks the boxes as they are completed.

- [ ] **Summary.** Write a summary of the changes. Explain what you did
to fix the issue, and why you did it. Present the changes in a logical
order. Instead of writing a summary in the pull request, you may push a
clean Git history.
- [ ] **Documentation.** Write documentation for new functions. Update
documentation for functions that changed, or complete documentation
where it is missing.
- [ ] **Tests.** Write unit tests for every change. Write the unit tests
that were missing before the changes. Include any examples from the
reported issue as integration tests.
- [ ] **Clean up.** The changes are already clean. Clean up anything
near the changes that you noticed while working. This does not mean only
spatially near the changes, but logically near: any code that interacts
with the changes!

---------

Co-authored-by: github-actions <github-actions@github.com>