remove parsing of \left-assoc and \right-assoc of \and and \or (#3675)
Blocked on https://github.com/runtimeverification/k/pull/3679

This completes the changes to the booster with respect to multi-ary \and
and \or. Once the \left-assoc and \right-assoc syntax are no longer
being emitted by the frontend, we can remove them from the backends.

### Scope:

### Estimate:

---

###### Review checklist

The author performs the actions on the checklist. The reviewer evaluates
the work and checks the boxes as they are completed.

- [ ] **Summary.** Write a summary of the changes. Explain what you did
to fix the issue, and why you did it. Present the changes in a logical
order. Instead of writing a summary in the pull request, you may push a
clean Git history.
- [ ] **Documentation.** Write documentation for new functions. Update
documentation for functions that changed, or complete documentation
where it is missing.
- [ ] **Tests.** Write unit tests for every change. Write the unit tests
that were missing before the changes. Include any examples from the
reported issue as integration tests.
- [ ] **Clean up.** The changes are already clean. Clean up anything
near the changes that you noticed while working. This does not mean only
spatially near the changes, but logically near: any code that interacts
with the changes!

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>