tarball-compare tool + store constraints in a set (#349)
## Store constraints  as `Set`s
The `constraints` field n `Pattern` as well as `requires` and `ensures`
fields in `RewriteRule` now hold a `Set Predicate` to avoid ordering
issues and duplication.

## `tarball-compare` tool

Closes #325 

Tool to compare two bug report tarballs (or directories containing an
unpacked bug report) for correspondence.
The following comparisons are performed:

* compare number of request/response files in the tarball
* compare definition.kore file in the tarball (top-level)
* Then, for each pair of requests or responses in the rpc_* directories:
  * compare file contents and file size of json files
  * compare the number of steps (depth) in responses to execute requests
* internalise states in response data (based on the definition.kore
file) and compare internal data (displaying diff if not equal)

The tool can work on:

* a single gzipped tarball containing other tarballs for many tests
(when given a single argument)
* two tarballs or directories of unpacked bug report data, assumed to
contain the same `rpc_*` directory names (when given two arguments, each
one either a tar file or a directory)

---------

Co-authored-by: Sam Balco <goodlyrottenapple@gmail.com>