Allow forcing pattern-wide simplfiication in `kore-rpc-booster` (#345)
This PR modifies add an option to `kore-rpc-booster` to simplify the
state in Kore every fixed number of steps. This allows to mitigate
growth of terms in Booster caused by a sub-expression that Booster is
unable to simplify. This is an alternative to the forced fallback to
Kore implemented by
https://github.com/runtimeverification/hs-backend-booster/pull/340.

I've added a run of `test-foundry-bug-report` integration tests over
`kore-rpc-booster --interim-simplification 100`.

Other changes include:
* new custom level `-l Depth` that prints the current depth before every
rewrite step
* new utilities to calculate various integral metrics of terms: size in
bytes, number of symbol application and number of variables. These are
useful when emitted via the above `Depth` log-level, but that is
semantic-specific, hence I'm not adding the actual use cases here.

---------

Co-authored-by: Sam Balco <goodlyrottenapple@gmail.com>