Update `generate-regression-tests.sh` work with modern evm-semantics (#3687)
Fixes #3686

This PR does the following
- Updates `generate-regression-tests.sh` to work with modern versions of
evm-semantics which no longer use a Makefile-based build system.
- Fixes all issues reported by `shellcheck`.
- Runs the new script and updates all the `regression-evm` tests.

---

###### Review checklist

The author performs the actions on the checklist. The reviewer evaluates
the work and checks the boxes as they are completed.

- [ ] **Summary.** Write a summary of the changes. Explain what you did
to fix the issue, and why you did it. Present the changes in a logical
order. Instead of writing a summary in the pull request, you may push a
clean Git history.
- [ ] **Documentation.** Write documentation for new functions. Update
documentation for functions that changed, or complete documentation
where it is missing.
- [ ] **Tests.** Write unit tests for every change. Write the unit tests
that were missing before the changes. Include any examples from the
reported issue as integration tests.
- [ ] **Clean up.** The changes are already clean. Clean up anything
near the changes that you noticed while working. This does not mean only
spatially near the changes, but logically near: any code that interacts
with the changes!

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>