Serialize the unknown predicate as JSON on `DecidePredicateUnknown` (#3694)
This PR changes the JSON RPC handler for `DecidePredicateUnknown`.

Before we pretty-printed (ugly-printed, let's be honest) the predicate
and returned that as the "data" field of the JSON RPC error. Now we
conjunct the predicates and side conditions, serialize the conjunction
as JSON and return that as "data".

This will solve two sources of paper cuts when using Kontrol:
* the server won't spam `stderr` with incomprehensible Kore output when
encountering difficult predicates
* `pyk`/`kontrol` can handle server errors with `"code": 5,
"message":"Smt solver error` better, for example by pretty-printing the
JSON predicate into surface K or tool-specific representation.