Add optional unknown predicate into "execute" response (#3697)
Add `"unknown-predicate"` as an optional field to `"execute"` JSON-RPC
response. The intention is to change servers' response on encountering
`ErrorDecidePredicateUnknwon`, i.e. when Z3 returns Unknwon while
checking a rewrite rule's side condition.

The current behavior is to return `SmtSolverError` even if progress was
made. I.e. we could have made 100 rewrite steps and only get Unknown
during step 101.

Instead, we should return the state before the error, and the predicate
that Z3 had trouble deciding. This way, the client will be able to make
an informed decision about that to do next and/or report the problematic
predicate to the user, so that they can add lemmas if appropriate.

Unfortunately, altering `kore-rpc` itself is quite difficult. therefore
I will implement a prototype of this in `Proxy.hs` of
`kore-rpc-booster`. The type and docs need to be update here though.

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>