Revert "Do not simplify before calling kore-rpc #362" (#403)
Revert #362 ~and #377~

* #362 causes Kore to return spurious branches when branching at depth
0. It looks like we must simplify the fallback state, because Kore will
not do it in this case. Alternatively, we need to modify Kore to call
the simplifier in the input state.
* ~#377 causes #401. In the state that is simplified to Bottom by
`kore-rpc-booster` there's a number of `==K` between `Bytes` functions
and `Bytes` domain value. It may be that our new `==K` simplifier on
Booster is not doing the right thing. It also may be that it does! Needs
further investigation.~ this has been fixed in #404

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>