Ceil simplifier + static ceil computation for rewrite rules (#402)
This PR introduces preliminary work on allowing the booster to apply
rewrite rules with partial function symbols in the requires/RHS of the
rule. We know the LHS does not contain function symbols and assume (but
should eventually check) that the given configuration is defined, i.e.
any partial functions are defined for the given input.

Up until now, we simply disallowed the application of any rules where
the RHS contained partial function symbols and this PR restricts this
further, by disallowing partial symbols in the requires clause. This is
because now that we are using an SMT solver to check whether the current
configuration conditions/predicates `C` imply the requires clause `R` of
a given rule (i.e. `C => R` is valid), we should really first check that
`C => #Ceil ( R )` is valid and only then check `C /\ #Ceil ( R ) => R`,
as the SMT solver assumes the totality of all functions. See
https://github.com/runtimeverification/haskell-backend/issues/3603 for
an example where this matters.

## Static ceil computation for rewrite rules

To deal with ceil, this PR introduces a static check which recursively
computes the `#Ceil` conditions of each rule. That is, for every rule:

1) We first compute `ceil ( LHS )`, which takes a term and returns a
list/set of predicates. This is done in the obvious way:
* For constructors and functions, we simply recurse into the arguments
to compute and merge `ceil`'s of all the children.
* For domain values, `#Ceil` is always `Top`, so we return an empty set.
* For any partial functions `f(a1, ..., an)`, we recursively compute
`ceil(a1) ++ ... ++ ceil(a_n)` of the children as well as symbolically
executing `#Ceil (f(a1, ..., an))`, given the user supplied `#Ceil`
rules found in the current definition and add the evaluated predicate to
the set
        * Finally, for any variable `X` we return `[#Ceil(X)]`
2) We similarly compute `ceil ( RHS )` and `ceil ( requires )`
3) We now compute the set difference `ceil ( RHS ) \\ ceil ( LHS )` and
`ceil ( requires ) \\ ceil ( LHS )`
4) Finally:
* If both sets are empty, we can conclude that the rule does in fact
preserve definedness and mark the rule as such
* If both sets contain only predicates which are concrete, we can call
the LLVM backend to evaluate them and mark the rule as
definedness-preserving if none evaluate to false.

## Future work

Given the static check above, there are several other cases to be
considered:

* Both sets contain only predicates (i.e. no `#Ceil( ... )`) in which
case, we should add a new set of definiteness predicates `D` to the rule
and when applying a rule check that the current configuration
conditions/predicates `C` implies `D`, `C /\ D` implies the requires `R`
clause and if the rule application is successful, merge `D`,`R` and `C`
into the new set of constraints for the new state
* Either set contains one or more `#Ceil( f( ... ) )` predicates,
indicating that the simplifier hasn't been able to find a ceil rule for
the given symbol. Here, we have two choices. We could enforce this
static check and continue to refuse to evaluate any such rules, making
the writer of the semantics change the rules or write the ceil
conditions for all partial functions. However, I'm not sure how feasible
this is. Another option could be to do what (I think) the old backend
does and allow the application of such rule and then try to simplify any
partial functions in both the requires clause and the RHS until it
either gets stuck or rewrites the partial functions to total ones. This
option is probably necessary, because if we start to use the SMT solver
to evaluate side-conditions in function applications, any partial
functions there will need their ceil computed and sent to the solver
first to check for definedness for the given arguments. This could
potentially get rather messy :/

### Static ceil computation for partial functions/simplifications

As discussed above, if we want to send partial functions to SMT or want
to start being able to apply partial functions during e.g. the
simplification phase, we will also likely need to either pre-compute the
ceil conditions statically or dynamically keep rewriting an checking
ceil until we hit a point in function application where the `#Ceil`s
disappear.


Fixes #318