384 409 rework predicate internalisation (#416)
Predicates are now always internalised as a group, and classified into
different kinds of predicates.

1. Input patterns are split up at top-level `\and`s.
2. predicate parts are classified into 
   * `BoolPred`icate (containing a term of `Bool` sort),
   * `CeilPred`icate, containing `\ceil(term)`
* `SubstitutionPred`icate, containing an equation `\equals(var, term)`
with a single variable on one side that is not present (i.e. free) on
the other side
* `UnsupportedPred`icate, which is ignored, left untouched, and re-added
to the output
3. the set of `SubstitutionPred`icates is vetted to ensure it represents
a valid non-circular and unambiguous substitution
* Only one equation per variable is allowed (otherwise all those
equations become `BoolPred`icates).
* There must not be a circular dependency between any variables in the
equations

it is an error to have any unsupported predicates in any part of a
definition.

Fixes #409 
Closes #384

---------

Co-authored-by: github-actions <github-actions@github.com>