Handle conflicting `DomainValue`s in `bindVariable` (#420)
In #408, we've made conflicting variable assignment during unification
indeterminate rather than failing. However, we can easily handle the
case of conflicting domain values.

The motivation for this is the `foundry.prank` rule in Kontrol, which
has a higher priority that the opcode fetching rule, and will be
attempted at every EVM step, if `<prank>` is set to `true`. This will
lead to Booster aborting every EVM step due to a conflicting assignment
to the `CD` variable in the `foundry.prank` rule.