Concurrent `kore-rpc` server (#3704)
Implements the proposal from
https://github.com/runtimeverification/hs-backend-booster/issues/383

Namely:

`add-module`
-------------

* now returns the id of a given module, of the form
`m<sha256_of_given_module>`
* computes the SHA256 hash of the unparsed module string and saves the
internalised module under this key
* takes an optional `name-as-id: true` argument which implements the
previous behaviour of `add-module` (still returns the new id), i.e. adds
the module to the module map under the module name as well as the id.
* if the same module is added twice with `name-as-id: true`, the second
request will fail with a `Duplicate module` error
* if the same module is sent twice with `name-as-id: false` or without
`name-as-id`, the second request is idempotent and will succeed

As discussed on the issue, the IDs and original names are not disjoint
for ease of implementation. the `m` pre-pended to the hash is necessary
to make the name a valid kore identifier.

I have also not added the `modules` field to other methods as I'm unsure
whether this should be replacing the current `module` parameter, which
would be a backwards incompatible change to the API that would break a
lot of our integration tests, or if it should be an additional field,
@tothtamas28?

---------

Co-authored-by: github-actions <github-actions@github.com>