Assume definedness of current configuration before rewriting (#3705)
Fixes https://github.com/runtimeverification/haskell-backend/issues/3703

Kore is designed to compute and check definedness of the current
configuration before attempting rewrite rules. In cases of large
concrete maps and sets with a handful of symbolic variables, the created
massive `#Ceil` predicates grind Kore to a halt.

However, when falling back to Kore from Booster, everything must be
defined, though it is difficult to convey this information from Booster
to Kore.

As an experimental workaround, we assume definiteness of every sub-term
of the current configuration before attempting rewrite rules. This
information should probably be conveyed from Booster in future.

---------

Co-authored-by: github-actions <github-actions@github.com>