Enable preserving-definedness atrribute for rules discovered during ceil analysis (#426)
Closes #424.

This is a fix for the ceil analysis introduced in #402. We successfully
infer definedness, but don't currently mark the rules as preserving
definedness correctly.

The added integration tests makes sure that a rule that contains a
partial function on the RHS (discovered to be preserving definedness
during the ceil analysis) will apply during execution.

---------

Co-authored-by: Jost Berthold <jost.berthold@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>