(Dev.) Options to control fall-back reasons and post-exec simplification (#436)
Fixes #434 

* `--fallback-on REASON1,REASON2,...` allows users to set a custom set
of reasons on which a fall-back to kore-rpc will be performed. The
default is `Branching,Stuck,Aborted` (behaviour as before).
* `--no-post-exec-simplify` switches off the post-exec simplification
with kore-rpc (booster results will always be simplified with booster's
simplifier before returning).

Both options are meant for developer experimentation. In particular, it
is probably not sensible to not have `Aborted` in the `fallback-on` set
of halt reasons.

---------

Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Georgy Lukyanov <mail@geo2a.info>