Revert "Concurrent `kore-rpc` server (#3704)" (#3707)
Reverts
https://github.com/runtimeverification/haskell-backend/pull/3704.

We would like to update the `haskell-backend` version used in
[hs-backend-booster](https://github.com/runtimeverification/hs-backend-booster/)
in order to merge
https://github.com/runtimeverification/hs-backend-booster/pull/421. We
cannot have
https://github.com/runtimeverification/haskell-backend/pull/3704
(concurrent `kore-rpc`), because we will need to merge
https://github.com/runtimeverification/hs-backend-booster/pull/427
(concurrent `kore-rpc-booster`). We do not want to merge the concurrent
`kore-rpc-booster` PR before the Holiday break, since it is a braking
change for `pyk`.

The reverted PR will need to be reopened. I think it is best to merge it
in sync with
https://github.com/runtimeverification/hs-backend-booster/pull/427,
preparing a relevant change to `pyk` in advance.