Refactor equals k to follow hook structure (#458)
Behaviour has changed:
If the hook cannot evaluate the entire expression, the _original_ term
will be returned. Previously, a simplified version would be returned.
For instance, for a simplification of expression `Constr(A, 42) ==K
Constr(B, 42)`, the old version returned `A ==K B`, whereas the new
version returns the original term `Constr(A, 42) ==K Constr(B, 42)`.