Allow sending different log levels to different handles (#512)
This PR modifies the logging infrastructure by allowing to redirect logs
at different levels to different handles. The functionality is not yet
used and the previous behavior of logging everything to `stderr` is
preserved.

In a follow-up PR we may introduce logging of some of the levels to a
file. To achieve that, we will need to modify the `logLevelToHandle`
function in `tools/booster/Server.hs` (for `kore-rpc-booster`) and in
`dev-tools/booster-dev/Server.hs` (for `booster-dev`).