Move Booster's `runServer` to `booster-dev` (#514)
The `Booster.JsonRpc.runServer` function has it's only call site in the
`booster-dev` executable's `main`. I'm moving it there for consistency
with `kore-rpc-booster`, which has similar code in `Proxy.hs`.

My primary motivation is to have the calls to `runHandleLoggingT` be in
the executables' code, rather than leaking into Booster's library. It is
going to be easier to set-up logging options with this layout.