Prune the remainder branch early if its condition is unsatisfiable (#3745)
Fixes #3738

**TL;DR** prune the rule remainder early with a direct call to the SMT
solver on the remainder predicate, rather then constructing the
remainder configuration and simplifying it.

----------------------

With this change, upon computing a remainder condition after applying
rewrite rules, we immediately ask the SMT solver to decide the
satidfiability of this condition. If the solver returns unsat, we prune
the remainder branch.

Previously, we would add the remainder predicate to the path condition
and construct the remainder branch from the current configuration. We
will then call the simplifier to evaluate functions in the
configuration. It turns out, we may start evaluating functions under a
false assumption, leading to pathological cases of applying
simplifications which are invalid (see #3738).

---------

Co-authored-by: rv-jenkins <admin@runtimeverification.com>