479 do not backtranslate fancy sorts in get-model (#563)
Values of any sort other than `SortBool` and `SortInt` cannot currently
be back-translated from the SMT representation (all sorts unknown to the
solver are opaque). In particular, this is true for `Map`-sorted values
(which are used frequently in EVM).

* Do not attempt to back-translate values of sorts other than `SortBool`
and `SortInt`
* but still include the related SMT variables in the `get-values`
requestl;
* and indicate untranslated variables by adding `VAR = VAR` to the
returned substitution.

As shown in the added integration test, there are cases where the
`Map`-sorted value is relevant, but cannot currently be returned. We
might consider modeling `Map`s specifically as arrays in the SMT solver
in the future.

Fixes https://github.com/runtimeverification/kontrol/issues/479