Merge Match/Unify modules (#3810)
Fixes #3785 by merging the two versions of our matching algorithm, one
used in matching rewrite rules and another matching
function/simplification rules. This is a first pass which mostly just
merges the two algorithms, prserving most of the divergent behaviour
under the `Rewrite`/`Eval` `MatchType` flag. A followup refactor should
clean up this code further.

This PR is a port of
https://github.com/runtimeverification/hs-backend-booster/pull/573 where
the development and testing happened.

---------

Co-authored-by: github-actions <github-actions@github.com>